[BACKEND] Fixed seeds; fixed new routes and tests to adhere to authentication

+++ backend/services/employerService.js
@@ -2,6 +2,7 @@ const employerDAO = require('../DAO/employerDAO');
 
 const SuperError = require('../errors').SuperError;
 const ERR_BAD_REQUEST = require('../errors').ERR_BAD_REQUEST;
+const ERR_FORBIDDEN = require('../errors').ERR_FORBIDDEN;
 
 const locationService = require('../services/locationService');
 const domainOfActivityService = require('../services/domainOfActivityService');
@@ -23,11 +24,14 @@ module.exports = employerService = {
     getEmployerByPlacementId: (placementId) => {
         return employerDAO.getEmployerByPlacementId(placementId);
     },
-    createNewEmployer: async (details) => {
+    createNewEmployer: async (details, auth) => {
         if(details && typeof(details.location.country) == 'string' && typeof(details.location.city) == 'string' && typeof(details.name) == 'string' && typeof(details.description) == 'string' && !isNaN(details.domainOfActivityId)){
+            if(auth.userType !== 'EMPLOYER') {
+                throw new SuperError(ERR_FORBIDDEN, 'You cannot create an employer profile from a non-employer account.')
+            }
             let check = await domainOfActivityService.existsDomainOfActivity(details.domainOfActivityId);
             if(check){
-                let newEmployer = await employerDAO.addNewEmployer(details);
+                let newEmployer = await employerDAO.addNewEmployer(details, auth.id);
                 if(details.location){
                     newEmployer.location = await employerService.saveEmployerLocation(newEmployer.id, details.location);
                 }
@@ -49,7 +53,10 @@ module.exports = employerService = {
         return location;
     },
 
-    deleteEmployerById: (id) => {
-        return employerDAO.deleteEmployerById(id);
+    deleteEmployerById: async (id, auth) => {
+        if(auth.employerId !== id) {
+            throw new SuperError(ERR_FORBIDDEN, 'You cannot delete this employer profile');
+        }
+        return await employerDAO.deleteEmployerById(id);
     }
 };
